Separate the meeting date from recurring meeting titles
Instead of creating a different title with the date for each recurring
meeting, show the start and end date separately. This gives us space to
show both and the dates update when the rules update.
